% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{D}
\alias{D}
\title{Consecutive disparity index (D)}
\usage{
D(P)
}
\arguments{
\item{P}{a numeric vector}
}
\value{
a numeric value
}
\description{
Calculate D of the series P.
}
\examples{
\donttest{
P = c(2,3,4,5)
D(P)
}
}
