% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{PV}
\alias{PV}
\title{Proportional Variability index (PV)}
\usage{
PV(Z)
}
\arguments{
\item{Z}{a numeric vector}
}
\value{
a numeric value
}
\description{
Calculate PV of the variable Z
}
\examples{
\donttest{
Z = c(2,3,4,5)
PV(Z)
}
}
